+++ backend/src/services/categoryService.ts
